+++ fiat-api/fiat-api.gradle
@@ -17,6 +17,8 @@
 dependencies {
   api("com.netflix.spinnaker.kork:kork-plugins-api")
 
+  implementation "com.fasterxml.jackson.core:jackson-annotations"
+
   compileOnly("org.projectlombok:lombok")
   annotationProcessor("org.projectlombok:lombok")
 }
