+++ fiat-core/src/main/java/com/netflix/spinnaker/fiat/model/resources/Account.java
@@ -17,7 +17,8 @@
 package com.netflix.spinnaker.fiat.model.resources;
 
 import com.fasterxml.jackson.annotation.JsonIgnore;
-import com.netflix.spinnaker.fiat.model.Authorization;
+import com.netflix.spinnaker.fiat.api.Authorization;
+import com.netflix.spinnaker.fiat.api.Permissions;
 import java.util.Set;
 import java.util.stream.Collectors;
 import lombok.Data;
