+++ fiat-core/src/main/java/com/netflix/spinnaker/fiat/model/resources/Application.java
@@ -18,7 +18,8 @@ package com.netflix.spinnaker.fiat.model.resources;
 
 import com.fasterxml.jackson.annotation.JsonAnySetter;
 import com.fasterxml.jackson.annotation.JsonIgnore;
-import com.netflix.spinnaker.fiat.model.Authorization;
+import com.netflix.spinnaker.fiat.api.Authorization;
+import com.netflix.spinnaker.fiat.api.Permissions;
 import java.util.HashMap;
 import java.util.Map;
 import java.util.Set;
