+++ fiat-core/src/main/java/com/netflix/spinnaker/fiat/model/resources/Authorizable.java
@@ -16,7 +16,7 @@
 
 package com.netflix.spinnaker.fiat.model.resources;
 
-import com.netflix.spinnaker.fiat.model.Authorization;
+import com.netflix.spinnaker.fiat.api.Authorization;
 import java.util.Set;
 
 public interface Authorizable {
