+++ fiat-core/src/main/java/com/netflix/spinnaker/fiat/model/resources/BaseAccessControlled.java
@@ -16,7 +16,8 @@
 
 package com.netflix.spinnaker.fiat.model.resources;
 
-import com.netflix.spinnaker.fiat.model.Authorization;
+import com.netflix.spinnaker.fiat.api.Authorization;
+import com.netflix.spinnaker.fiat.api.Permissions;
 import java.util.List;
 import lombok.extern.slf4j.Slf4j;
