+++ fiat-core/src/test/groovy/com/netflix/spinnaker/fiat/model/resources/ResourceSpec.groovy
@@ -16,7 +16,8 @@
 
 package com.netflix.spinnaker.fiat.model.resources
 
-import com.netflix.spinnaker.fiat.model.Authorization
+import com.netflix.spinnaker.fiat.api.Authorization
+import com.netflix.spinnaker.fiat.api.Permissions
 import spock.lang.Specification
 
 class ResourceSpec extends Specification {
