+++ fiat-web/src/main/java/com/netflix/spinnaker/fiat/config/FiatConfig.java
@@ -3,7 +3,7 @@ package com.netflix.spinnaker.fiat.config;
 import com.google.common.collect.ImmutableList;
 import com.netflix.spectator.api.Registry;
 import com.netflix.spinnaker.config.PluginsAutoConfiguration;
-import com.netflix.spinnaker.fiat.model.Authorization;
+import com.netflix.spinnaker.fiat.api.Authorization;
 import com.netflix.spinnaker.fiat.model.resources.Application;
 import com.netflix.spinnaker.fiat.model.resources.Role;
 import com.netflix.spinnaker.fiat.permissions.DefaultFallbackPermissionsResolver;
